fix: stabilize reader progress flush timing
Prioritize flush-triggered page progress dispatch with userInitiated task priority so close-reader flushes are not delayed by background scheduling.

Wait for progress dispatch to settle based on pending/debounce/in-flight queues instead of in-flight tracking only, and use this in dashboard refresh and visited-item sync close paths.

Remove ReaderProgressTracker since dispatch service now owns the relevant wait-state checks.
